:building_construction: Extract WebSocket-related code into a separate file
