+++ midi-script/Socket.py
@@ -1,91 +1,30 @@
 from __future__ import absolute_import
 import socket
 import json
-import struct
-import sys
-import hashlib
-import base64
 import threading
 import time
 
 from .Config import PLUGIN_NAME, WEBSOCKET_HOST, WEBSOCKET_PORT
 from .Logging import logger
+from .WebSocket import (
+    OPCODE_CLOSE,
+    OPCODE_CONTINUATION,
+    OPCODE_PING,
+    OPCODE_PONG,
+    OPCODE_TEXT,
+    encode_close_frame,
+    encode_pong_frame,
+    encode_text_frame,
+    handshake,
+    to_bytes,
+    to_text,
+    try_read_frame,
+)
 
 try:
     import Queue as queue
 except ImportError:
     import queue
-
-
-WS_GUID = "258EAFA5-E914-47DA-95CA-C5AB0DC85B11"
-
-
-def _byte_at(data, index):
-    value = data[index]
-    if isinstance(value, int):
-        return value
-    return ord(value)
-
-
-def _to_bytes(data):
-    if data is None:
-        return b""
-    if isinstance(data, bytearray):
-        return bytes(data)
-    if sys.version_info[0] >= 3:
-        if isinstance(data, bytes):
-            return data
-        return data.encode("utf-8")
-    if isinstance(data, unicode):
-        return data.encode("utf-8")
-    return data
-
-
-def _to_text(data):
-    if sys.version_info[0] >= 3:
-        if isinstance(data, bytes):
-            return data.decode("utf-8")
-        return data
-    if isinstance(data, unicode):
-        return data
-    return data.decode("utf-8")
-
-
-def encode_text_frame(payload_bytes):
-    payload = _to_bytes(payload_bytes)
-    header = bytearray()
-    header.append(0x81)
-    length = len(payload)
-    if length < 126:
-        header.append(length)
-    elif length < 65536:
-        header.append(126)
-        header.extend(struct.pack("!H", length))
-    else:
-        header.append(127)
-        header.extend(struct.pack("!Q", length))
-    return bytes(header) + payload
-
-
-def encode_pong_frame(payload_bytes):
-    payload = _to_bytes(payload_bytes)
-    header = bytearray()
-    header.append(0x8A)
-    length = len(payload)
-    if length < 126:
-        header.append(length)
-    elif length < 65536:
-        header.append(126)
-        header.extend(struct.pack("!H", length))
-    else:
-        header.append(127)
-        header.extend(struct.pack("!Q", length))
-    return bytes(header) + payload
-
-
-def encode_close_frame():
-    return b"\x88\x00"
-
 
 # Send small frames on the caller thread (usually Live's MIDI thread). A
 # dedicated send thread adds ~10ms wake-up on macOS; keep it only for large
@@ -206,7 +145,7 @@ class Socket(object):
                 default=jsonReplace,
                 ensure_ascii=False,
             )
-            frame = encode_text_frame(_to_bytes(data))
+            frame = encode_text_frame(to_bytes(data))
             self._send_frame(frame)
         except Exception as e:
             logger.error("Error " + name + "(" + str(uuid) + "):")
@@ -319,7 +258,7 @@ class Socket(object):
             thread.start()
 
     def _handle_connection(self, conn):
-        leftover = self._handshake(conn)
+        leftover = handshake(conn)
         if leftover is None:
             try:
                 conn.close()
@@ -333,7 +272,7 @@ class Socket(object):
 
         self._send_frame(
             encode_text_frame(
-                _to_bytes(
+                to_bytes(
                     json.dumps(
                         {
                             "event": "connect",
@@ -347,7 +286,7 @@ class Socket(object):
             client,
         )
 
-        buffer = bytearray(_to_bytes(leftover))
+        buffer = bytearray(to_bytes(leftover))
         fragments = bytearray()
         fragment_opcode = None
 
@@ -359,24 +298,24 @@ class Socket(object):
                     break
                 if not data:
                     break
-                buffer.extend(bytearray(_to_bytes(data)))
+                buffer.extend(bytearray(to_bytes(data)))
 
                 while True:
-                    frame = self._try_read_frame(buffer)
+                    frame = try_read_frame(buffer)
                     if frame is None:
                         break
                     opcode, fin, payload = frame
 
-                    if opcode == 0x8:
+                    if opcode == OPCODE_CLOSE:
                         client.enqueue(encode_close_frame())
                         return
-                    if opcode == 0x9:
+                    if opcode == OPCODE_PING:
                         self._send_frame(encode_pong_frame(payload), client)
                         continue
-                    if opcode == 0xA:
+                    if opcode == OPCODE_PONG:
                         continue
 
-                    if opcode == 0x0:
+                    if opcode == OPCODE_CONTINUATION:
                         if fragment_opcode is None:
                             continue
                         fragments.extend(payload)
@@ -401,102 +340,13 @@ class Socket(object):
             logger.info("Client disconnected")
 
     def _handle_payload(self, opcode, payload):
-        if opcode != 0x1:
+        if opcode != OPCODE_TEXT:
             return
         try:
-            text = _to_text(bytes(payload))
+            text = to_text(bytes(payload))
             parsed = json.loads(text)
         except Exception as e:
             logger.error("Error decoding request:")
             logger.exception(e)
             return
         self._queue.put(parsed)
-
-    def _handshake(self, conn):
-        data = b""
-        while b"\r\n\r\n" not in data:
-            try:
-                chunk = conn.recv(4096)
-            except socket.error:
-                return None
-            if not chunk:
-                return None
-            data += _to_bytes(chunk)
-            if len(data) > 16384:
-                return None
-
-        header_blob, leftover = data.split(b"\r\n\r\n", 1)
-        try:
-            header_text = _to_text(header_blob)
-        except Exception:
-            return None
-
-        headers = {}
-        lines = header_text.split("\r\n")
-        for line in lines[1:]:
-            if not line or ": " not in line:
-                continue
-            key, value = line.split(": ", 1)
-            headers[key.strip().lower()] = value.strip()
-
-        key = headers.get("sec-websocket-key")
-        if not key:
-            return None
-
-        accept_src = _to_bytes(key + WS_GUID)
-        accept = base64.b64encode(hashlib.sha1(accept_src).digest())
-        if sys.version_info[0] >= 3:
-            accept = accept.decode("ascii")
-
-        response = (
-            "HTTP/1.1 101 Switching Protocols\r\n"
-            "Upgrade: websocket\r\n"
-            "Connection: Upgrade\r\n"
-            "Sec-WebSocket-Accept: " + accept + "\r\n\r\n"
-        )
-        try:
-            conn.sendall(_to_bytes(response))
-        except Exception:
-            return None
-        return leftover
-
-    def _try_read_frame(self, buffer):
-        if len(buffer) < 2:
-            return None
-
-        b1 = _byte_at(buffer, 0)
-        b2 = _byte_at(buffer, 1)
-        opcode = b1 & 0x0F
-        masked = (b2 & 0x80) != 0
-        length = b2 & 0x7F
-        index = 2
-
-        if length == 126:
-            if len(buffer) < 4:
-                return None
-            length = struct.unpack("!H", bytes(buffer[2:4]))[0]
-            index = 4
-        elif length == 127:
-            if len(buffer) < 10:
-                return None
-            length = struct.unpack("!Q", bytes(buffer[2:10]))[0]
-            index = 10
-
-        mask = None
-        if masked:
-            if len(buffer) < index + 4:
-                return None
-            mask = buffer[index : index + 4]
-            index += 4
-
-        if len(buffer) < index + length:
-            return None
-
-        payload = bytearray(buffer[index : index + length])
-        if mask is not None:
-            for i in range(len(payload)):
-                payload[i] = payload[i] ^ _byte_at(mask, i % 4)
-
-        del buffer[: index + length]
-        fin = (b1 & 0x80) != 0
-        return opcode, fin, payload
